+++ tests/test_parameters.py
@@ -455,7 +455,7 @@ def test_integrated_show_params_option(invoke, create_config):
             "✘",
             "✓",
             "SHOW_PARAMS_CLI_VERBOSITY",
-            "WARNING",
+            "<LogLevel.WARNING: 30>",
             "DeBuG",
             "COMMANDLINE",
         ),